Enqueue 0.4.6

+++ Frameworks/EMKeychainItem/EMKeychainItem.m
@@ -243,7 +243,7 @@
 	SecKeychainAttribute attributes[1];
 	attributes[0].tag = kSecProtocolItemAttr;
 	attributes[0].length = sizeof(newProtocol);
-	attributes[0].data = (void *)newProtocol;
+	attributes[0].data = (void *)(NSInteger)newProtocol;
 	
 	SecKeychainAttributeList list;
 	list.count = 1;
